+++ crates/cef-ui-util/src/lib.rs
@@ -1,7 +1,11 @@
-mod builds;
+mod download;
+mod link;
+mod paths;
 mod tools;
 mod util;
 
-pub use builds::*;
+pub use download::*;
+pub use link::*;
+pub use paths::*;
 pub use tools::*;
 pub use util::*;
