+++ crates/cef-ui-util/src/tools/mod.rs
@@ -1,7 +1,9 @@
 mod cef_artifacts;
 mod cef_build;
 mod cef_clean;
+mod macos;
 
 pub use cef_artifacts::*;
 pub use cef_build::*;
 pub use cef_clean::*;
+pub use macos::*;
